FixAll

+++ tasks/seq/golovkin_rowwise_matrix_partitioning/perf_tests/main.cpp
@@ -52,7 +52,7 @@ TEST(golovkin_rowwise_matrix_partitioning, test_pipeline_run) {
   ppc::core::Perf::print_perf_statistic(perfResults);
 
   ASSERT_EQ(result.size(), static_cast<size_t>(N));
-  ASSERT_EQ(result[0].size(), N);
+  ASSERT_EQ(result[0].size(), static_cast<size_t>(N));
 }
 
 TEST(golovkin_rowwise_matrix_partitioning, test_task_run) {
@@ -92,5 +92,5 @@ TEST(golovkin_rowwise_matrix_partitioning, test_task_run) {
   ppc::core::Perf::print_perf_statistic(perfResults);
 
   ASSERT_EQ(result.size(), static_cast<size_t>(N));
-  ASSERT_EQ(result[0].size(), N);
+  ASSERT_EQ(result[0].size(), static_cast<size_t>(N));
 }
